move batch changes methods to their own files
