+++ vinyldns/api.go
@@ -11,45 +11,3 @@ limitations under the License.
 */
 
 package vinyldns
-
-import (
-	"encoding/json"
-)
-
-// BatchRecordChanges returns the list of batch record changes
-func (c *Client) BatchRecordChanges() ([]RecordChange, error) {
-	changes := &BatchRecordChanges{}
-	err := resourceRequest(c, batchRecordChangesEP(c), "GET", nil, changes)
-	if err != nil {
-		return nil, err
-	}
-
-	return changes.BatchChanges, nil
-}
-
-// BatchRecordChange returns the batch record change
-// associated with the change whose ID it's passed.
-func (c *Client) BatchRecordChange(changeID string) (*BatchRecordChange, error) {
-	change := &BatchRecordChange{}
-	err := resourceRequest(c, batchRecordChangeEP(c, changeID), "GET", nil, change)
-	if err != nil {
-		return nil, err
-	}
-
-	return change, nil
-}
-
-// BatchRecordChangeCreate creates the batch record change it's passed.
-func (c *Client) BatchRecordChangeCreate(change *BatchRecordChange) (*BatchRecordChangeUpdateResponse, error) {
-	cJSON, err := json.Marshal(change)
-	if err != nil {
-		return nil, err
-	}
-	var resource = &BatchRecordChangeUpdateResponse{}
-	err = resourceRequest(c, batchRecordChangesEP(c), "POST", cJSON, resource)
-	if err != nil {
-		return &BatchRecordChangeUpdateResponse{}, err
-	}
-
-	return resource, nil
-}
